+++ src/reusables.js
@@ -1,4 +1,5 @@
 export function asArray(x) {
+    if (!x) return [];
     return Array.isArray(x) ? x : [x];
 }